requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
python-dotenv==1.0.0
PyMySQL==1.1.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.1.1
email-validator==2.0.0
Flask==2.*
Flask-SQLAlchemy==3.*
Flask-Login==0.6.*
Flask-Bcrypt==1.*
Werkzeug==2.*